Analysis of 100 development set abstracts full workflow.
To understand why our performance on the official evaluation dataset was
so poor, we processed another 100 abstracts from the development set
with the full workflow (NER step and CID step) to determine performance.
The original 200 abstract development set did not test the NER step and
also had cheaters, leading to an incomplete dataset.