Add filtering method for redundant MeSH relations.
The BioCreative annotation guidelines state that only the most specific
instance of a disease should be used in a CID relation. At the moment we
do not check for this redundancy.

mesh_filter.py implements a simple version of redundancy checking. A
disease is considered redundant if there is another disease which lies
directly below it in the MeSH hierarchy tree. MeSH tree numbers are used
to determine the hierarchy.