Remove analysis of title's effect on performance.
Removed the analysis of whether crowd was doing worse on sentences which
were titles compared to regular sentences. We found that if anything it
seemed that the crowd was doing better on relations when they were part
of a title. On closer examination it seemed that the effect was due more
to the number of tasks, and not from the titles themselves.