Complete revision of Relation class.
Simplified the relation generation and validation so that relations only
use pairs of single OntologyIDs, but will use all annotations to
generate the possible relation pairs. This way the data is consistent
with the standard given by the gold standard, but all annotations will
be highlighted for the crowd to verify.