Add analysis of single abstract-task relation verification.
Add analysis of CrowdFlower job 847908. In this job, we tested whether
the workers would perform any better when provided with the full context
of the entire abstract when making judgments about whether a relation
was true. This task showed a single relation pair with the entire
abstract, even if the two concepts co-occur within the same sentence.